ipv4/ipv6: route: Fix Clang 21 -Wdefault-const-init-var-unsafe warning
This commit resolves a compilation warning in `net/ipv6/route.c` when
building with Clang 21:
`default initialization of an object of type 'typeof (rt->dst.expires)' (aka 'const unsigned long') leaves the object uninitialized [-Werror,-Wdefault-const-init-var-unsafe]`

The warning occurs on line 403 where `time_after` is used with
`rt->dst.expires`. The `rt->dst.expires` field is a `const unsigned long`,
and Clang's `typeof` operator, used within the `time_after` macro,
identifies an implicit uninitialized variable creation.

To fix this, `rt->dst.expires` is now explicitly cast to `unsigned long`
before being passed to `time_after`. This ensures that the type checking
mechanism within the macro is satisfied and prevents the warning,
while maintaining the correct functionality.

+++ net/ipv6/route.c
@@ -388,7 +388,7 @@ static void ip6_dst_ifdown(struct dst_entry *dst, struct net_device *dev,
 static bool __rt6_check_expired(const struct rt6_info *rt)
 {
 	if (rt->rt6i_flags & RTF_EXPIRES)
-		return time_after(jiffies, rt->dst.expires);
+		return time_after(jiffies, (unsigned long)rt->dst.expires);
 	else
 		return false;
 }
@@ -400,7 +400,7 @@ static bool rt6_check_expired(const struct rt6_info *rt)
 	from = rcu_dereference(rt->from);
 
 	if (rt->rt6i_flags & RTF_EXPIRES) {
-		if (time_after(jiffies, rt->dst.expires))
+		if (time_after(jiffies, (unsigned long)rt->dst.expires))
 			return true;
 	} else if (from) {
 		return rt->dst.obsolete != DST_OBSOLETE_FORCE_CHK ||
